if you know, tell me.

+++ database/migrations/2024_05_19_111543_create_courses_table.php
@@ -18,7 +18,7 @@ return new class extends Migration {
             $table->string('image'); // it seems a little shorter :)
             $table->longText('description');
             $table->Decimal('cost'); // Just deleted the is_free attribute. If it is free, The cost is 0, and it can be handled by the frontDevs.
-            $table->integer('average_rating')->default(0); // from 1 star to 5.
+            $table->double('average_rating')->default(0.0); // from 1 star to 5.
             $table->boolean('is_reviewed')->default(false);
             $table->timestamps();
         });
